+++ Locales/enUS.lua
@@ -1826,6 +1826,7 @@ L["Texture & Colors"] = true
 L["Threshold Mode"] = true
 L["Timing"] = true
 L["Tint"] = true
+L["Tint Entire Bar"] = true
 L["Track Highest Duration"] = true
 L["Track Lowest Duration"] = true
 L["Trigger Mode"] = true